requirements.txt
# requirements.txt
google-adk
google-cloud-aiplatform
fastapi
uvicorn[standard] # Includes performance extras
requests
gunicorn # To manage uvicorn workers in production